chore: SCRUM-34 테스트 코드 패키지 구조 변경
