--- src/test/java/com/ixi_U/chatbot/constants/TestConstants.java
@@ -1,10 +1,0 @@
-package com.ixi_U.chatbot.constants;
-
-public class TestConstants {
-
-    public static final String CHATBOT_WELCOME_MESSAGE = """
-            안녕하세요, 고객님! 어떤 요금제를 찾고 계실까요?
-            관심있는 혜택 또는 원하는 조건을 말씀해주시면 최적의 요금제를 안내해드리겠습니다!
-            예시) "넷플릭스 있는 요금제 중 가장 싼 요금제가 뭐야?", "데이터 10기가 이상인 요금제 알려줘"
-            """;
-}
