+++ src/test/java/com/ixi_U/chatbot/testutil/TestDataFactory.java
@@ -15,10 +15,10 @@ public class TestDataFactory {
     public static GeneratePlanDescriptionRequest createValidPlan() {
 
         SingleBenefitDTO sbr = SingleBenefitDTO.create(
-                UUID.randomUUID().toString(), "테스트 단일 혜택", "기본 단일 혜택", BenefitType.DEVICE);
+                UUID.randomUUID().toString(), "테스트 단일 혜택", BenefitType.DEVICE);
 
         BundledBenefitDTO bbr = BundledBenefitDTO.create(
-                UUID.randomUUID().toString(), "테스트 번들 혜택", "기본 번들 혜택", 1, List.of(sbr));
+                UUID.randomUUID().toString(), "테스트 번들 혜택", 1, List.of(sbr));
 
         return GeneratePlanDescriptionRequest.create(UUID.randomUUID().toString(), "테스트 요금제", 100,
                 59000, List.of(bbr), List.of(sbr));
@@ -30,12 +30,12 @@ public class TestDataFactory {
     public static GeneratePlanDescriptionRequest createNullFieldPlan() {
 
         SingleBenefitDTO sbr1 = SingleBenefitDTO.create(
-                UUID.randomUUID().toString(), null, null, BenefitType.DEVICE);
+                UUID.randomUUID().toString(), null, BenefitType.DEVICE);
         SingleBenefitDTO sbr2 = SingleBenefitDTO.create(
-                UUID.randomUUID().toString(), null, null, BenefitType.DEVICE);
+                UUID.randomUUID().toString(), null,  BenefitType.DEVICE);
 
         BundledBenefitDTO bbr = BundledBenefitDTO.create(
-                UUID.randomUUID().toString(), "테스트 번들 혜택", "기본 번들 혜택", 1, List.of(sbr1));
+                UUID.randomUUID().toString(), "테스트 번들 혜택", 1, List.of(sbr1));
 
         return GeneratePlanDescriptionRequest.create(UUID.randomUUID().toString(), "테스트 요금제", 100,
                 59000, List.of(bbr), List.of(sbr2));
